Update print_area_bed_mesh.cfg
Added additional checks when using a Euclid probe.

Currently the script checks for the presence of a Klicky probe by looking for the _Probe_Variables macro. Euclid doesn't use this macro so the script doesn't attach the Euclid probe.

New check added to look for the EuclidProbe macro which is defined when using a Euclid probe. You need to be using v5 of the Euclid macros (https://github.com/nionio6915/Euclid_Probe/blob/main/Firmware_Examples/Klipper/00-euclid_exampleV5.cfg) so that the probe is deployed and stowed with the "DEPLOY_PROBE" and "STOW_PROBE" macros, which were M401 and M402 in earlier versions of Euclid Macros.

Tested as working on a Voron V2.4.